requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.6.3
async-lru==1.0.2
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
cachetools==4.2.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
DAWG-Python==0.7.2
docopt-ng==0.9.0
frozenlist==1.4.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.0
openai==0.28.1
packaging==23.2
pathy==0.10.2
preshed==3.0.9
pyaes==1.6.1
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pymorphy3==1.2.1
pymorphy3-dicts-ru==2.4.417150.4580142
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
python-socks==2.4.3
python-telegram-bot==13.7
pytz==2023.3
requests==2.31.0
rsa==4.9
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0-py3-none-any.whl
six==1.16.0
smart-open==6.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
Telethon==1.31.1
TgCrypto==1.2.5
thinc==8.2.1
tornado==6.3.3
tqdm==4.66.1
transliterate==1.10.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
wasabi==1.1.2
weasel==0.3.2
yarl==1.9.2
